% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{mrdivide}
\alias{mrdivide}
\title{Matlab mrdivide (/) operator}
\usage{
mrdivide(A, B, pinv = TRUE)
}
\description{
Solve systems of linear equations xA = B for x
x = B/A solves the system of linear equations x*A = B for x. The matrices A and B must contain the same number of columns. MATLAB® displays a warning message if A is badly scaled or nearly singular, but performs the calculation regardless.
If A is a scalar, then B/A is equivalent to B./A.
If A is a square n-by-n matrix and B is a matrix with n columns, then x = B/A is a solution to the equation x*A = B, if it exists.
If A is a rectangular m-by-n matrix with m ~= n, and B is a matrix with n columns, then x = B/A returns a least-squares solution of the system of equations x*A = B.
}
